+++ src/main/java/cart/dto/response/ResponseProductDto.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 public class ResponseProductDto {
