+++ src/main/java/cart/service/CartService.java
@@ -3,9 +3,9 @@ package cart.service;
 import cart.dao.ProductDao;
 import cart.dao.entity.ProductEntity;
 import cart.domain.Product;
-import cart.dto.RequestCreateProductDto;
-import cart.dto.RequestUpdateProductDto;
-import cart.dto.ResponseProductDto;
+import cart.dto.request.RequestCreateProductDto;
+import cart.dto.request.RequestUpdateProductDto;
+import cart.dto.response.ResponseProductDto;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
